// src/generated/LoginFlowsView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { SsoIdentityProviderView } from "./SsoIdentityProviderView";

export type LoginFlowsView = { password: boolean, oidc: boolean, oidc_registration: boolean, sso: boolean, oauth_aware_preferred: boolean, sso_identity_providers: Array<SsoIdentityProviderView>, };